src/requirements_production.txt
# Core Dependencies
google-genai>=0.2.0
google-adk>=0.1.0
Pillow>=10.0.0
pdf2image>=1.16.0

# Async and utilities
asyncio

# Optional: For production deployments
python-dotenv>=1.0.0

